Dashboard DTOs + repo + services in progress - Pt 4

+++ src/main/java/com/inventoryapp/demo/dtos/MonthToDateReportingDTO.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 @Data
 public class MonthToDateReportingDTO {
     private String shop;
-    private int salesNo;
+    private Long salesNo;
     private double salesTo;
     private double salesMg;
     private double salesMgAvg;
